# backend/backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: CityItHack Access Control API
  version: 2.1.0
  description: |
    Система контроля доступа с поддержкой:
    - Создание карт с ролями пользователей (admin, permanent, temporary, guest)
    - Автоматическая ротация ключей с нулевым простоем
    - QR коды для гостевых пропусков
    - Удаленная аттестация ридеров
    - Аппаратная защищенность (обертка ключей AES-GCM)
    - Полная система аудита и логирования событий
    - История сканирований и верификаций с временными метками
    - Мониторинг подозрительной активности и безопасности
    - Статистика и аналитика по всем операциям
    - Экспорт данных аудита в CSV формате
servers:
  - url: http://localhost:8080
    description: Local server
  - url: http://localhost:8081
    description: Development server
paths:
...
  /api/qr/verify:
    post:
      summary: Верификация QR кода
      operationId: verifyQrCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QrVerifyRequest'
      responses:
        '200':
          description: Результат проверки QR кода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrVerifyResponse'
        '400':
          description: Ошибка валидации QR кода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrVerifyResponse'
  /api/qr/generate/{cardId}:
    get:
      summary: Генерация QR кода для карты
      operationId: generateQrCode
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
          description: base64url(16) идентификатор карты
      responses:
        '200':
          description: QR код сгенерирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrGenerateResponse'
        '404':
          description: Карта не найдена
  /api/attest/challenge/{readerId}:
    post:
      summary: Генерация челленджа для аттестации ридера
      operationId: generateAttestationChallenge
      parameters:
        - in: path
          name: readerId
          required: true
          schema:
            type: string
          description: Идентификатор ридера
      responses:
        '200':
          description: Челлендж сгенерирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationChallenge'
  /api/attest/verify/{readerId}:
    post:
      summary: Верификация аттестации ридера
      operationId: verifyAttestation
      parameters:
        - in: path
          name: readerId
          required: true
          schema:
            type: string
          description: Идентификатор ридера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttestationVerifyRequest'
      responses:
        '200':
          description: Аттестация успешна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationResponse'
        '400':
          description: Ошибка аттестации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationResponse'
  /api/attest/status/{readerId}:
    get:
      summary: Статус аттестации ридера
      operationId: getAttestationStatus
      parameters:
        - in: path
          name: readerId
          required: true
          schema:
            type: string
          description: Идентификатор ридера
      responses:
        '200':
          description: Статус аттестации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationStatus'
...
components:
  schemas:
...
    QrVerifyRequest:
      type: object
      required: [qrCode]
      properties:
        qrCode:
          type: string
          description: Base64URL encoded QR code
    QrVerifyResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [OK, FAIL]
        cardId:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    QrGenerateResponse:
      type: object
      required: [status, cardId, qrCode]
      properties:
        status:
          type: string
          enum: [OK]
        cardId:
          type: string
        qrCode:
          type: string
        owner:
          type: string
        userRole:
          type: string
    AttestationChallenge:
      type: object
      required: [challenge, readerId]
      properties:
        challenge:
          type: string
          description: UUID challenge string
        readerId:
          type: string
    AttestationVerifyRequest:
      type: object
      required: [challenge, signature]
      properties:
        challenge:
          type: string
        signature:
          type: string
          description: Device signature
    AttestationResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [OK, FAIL]
        readerId:
          type: string
          nullable: true
        attestedAt:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    AttestationStatus:
      type: object
      required: [status, readerId]
      properties:
        status:
          type: string
          enum: [ATTESTED, NOT_ATTESTED]
        readerId:
          type: string
        attestedAt:
          type: string
          nullable: true
        ageMinutes:
          type: integer
          nullable: true
...